moved utility functions into dedicated file
