+++ docs/functions.md
@@ -158,44 +158,3 @@ So that the fitted value `x[i]` are
 ```python
 print(f"f({x[i]:.2e}) = {f(x[i], *params):.2e} (+{sigma_pos[i]:.2e}/-{sigma_neg[i]:.2e})")
 ```
-
-## Generate Probability for Sigma Interval
-
-```py
-get_sigma_probability(n: float = 1)
-```
-
-To get probability to fall into n-sigma interval call 
-
-| Parameters | | |
-|----------|----------|-----------------|
-| **Name** | **Type** | **Description** |
-| n        | float    | Number of sigmas in interval
-
-| Returns | | |
-|----------|----------|-----------------|
-| **Name** | **Type** | **Description** |
-| p        | float    | Probability of falling into sigma interval. $P(\mu - n*\sigma < X < \mu + n*\sigma) $
-
-
-## Generating Thresholds
-
-```py
-generate_thresholds(data, lower_frac=1/6, upper_frac=5/6)`
-```
-
-Given a bootstrapped distribution, generate a custom confidence interval.
-
-| Parameters | | |
-|----------|----------|-----------------|
-| **Name** | **Type** | **Description** |
-| data     | np.array | At a defined point x give the y value for all points in parameter space.
-| lower_frac | float  | Fraction of data below lower threshold
-| upper_frac | float  | Fraction of data below upper threshold
-
-
-| Returns | | |
-|----------|----------|-----------------|
-| **Name** | **Type** | **Description** |
-| lower_threshold | float | Point defined by `lower_frac` 
-| upper_threshold | float | Point defined by `upper_frac`
